revert rename misc services
